Add anti-aliasing:
Usage:

Toggle AA with the boolean in main.

The float used for the function is the weight distributed to the actual pixel rendered.

The lower the number, the more the pixels around will affect the pixel rendered. (1.0 is they are all treated equally, 4.0 is the pixel rendered is being counted 4x as much as pixels surrounding).